refactor: extract GetAdapterHandlerFromType and Id types to global-types for better type management
